requirements.txt
torch
numpy
tqdm
transformers
bs4
faiss-cpu
dvc
dvc-gdrive
sentencepiece
